And update header too...

+++ src/lib/examine_pe.h
@@ -50,7 +50,7 @@ const void *exm_pe_entry_point_get(const Exm_Pe *pe);
 
 const IMAGE_EXPORT_DIRECTORY *exm_pe_export_directory_get(const Exm_Pe *pe, DWORD *count);
 
-unsigned char exm_pe_export_directory_function_ordinal_get(const Exm_Pe *pe, const IMAGE_EXPORT_DIRECTORY *ed, DWORD idx, WORD *ordinal);
+unsigned char exm_pe_export_directory_function_ordinal_get(const Exm_Pe *pe, const IMAGE_EXPORT_DIRECTORY *ed, DWORD idx, DWORD *ordinal);
 
 const char *exm_pe_export_directory_function_name_get(const Exm_Pe *pe, const IMAGE_EXPORT_DIRECTORY *ed, DWORD idx);
